Release

+++ app/src/main/java/com/taskdoday/taskdoday/MainActivity.java
@@ -1,4 +1,4 @@
-package com.example.taskdoday;
+package com.taskdoday.taskdoday;
 
 import android.app.AlarmManager;
 import android.app.NotificationChannel;
@@ -56,8 +56,8 @@ import java.util.Date;
 import java.util.GregorianCalendar;
 import java.util.List;
 
-import static com.example.taskdoday.R.menu.menu_main;
-import static com.example.taskdoday.R.menu.menu_main2;
+import static com.taskdoday.taskdoday.R.menu.menu_main;
+import static com.taskdoday.taskdoday.R.menu.menu_main2;
 
 public class MainActivity extends AppCompatActivity {
     RecyclerView recyclerView;
@@ -135,6 +135,9 @@ public class MainActivity extends AppCompatActivity {
         setSwipes();
 
     }
+
+
+
     public void notificationSetup() throws JSONException {
 
         SharedPreferences prefs = getSharedPreferences("com.exmample.taskdoday", MODE_PRIVATE);
@@ -155,8 +158,6 @@ public class MainActivity extends AppCompatActivity {
             tempcal.set(Calendar.HOUR, hour);
             tempcal.set(Calendar.MINUTE, minute);
             tempcal.add(Calendar.DAY_OF_MONTH, -1);
-            System.out.println("DDDDDDDDDDDDDDDDDDDDDDDDDDDDD3");
-            System.out.println(tempcal.getTime());
 
             Intent notifyIntent = new Intent(this, MyReceiver.class);
             PendingIntent pendingIntent = PendingIntent.getBroadcast
@@ -188,7 +189,6 @@ public class MainActivity extends AppCompatActivity {
             done(view);
         }
         calendar.setTimeInMillis(calendarDialog.getGcalendar().getTimeInMillis());
-        System.out.println(calendar.getTime());
 
         setDate();
         if(getIsOld()){
@@ -286,7 +286,6 @@ public class MainActivity extends AppCompatActivity {
                         RelativeLayout notaskl = findViewById(R.id.notaskslayout);
                         notaskl.setVisibility(View.VISIBLE);
                     }
-                    System.out.println("ZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZ" + myDataset);
                     mAdapter = new MyAdapter(getApplicationContext(), myDataset, myStatus, myID, getIsOld(), true, darkMode);
                     recyclerView.setAdapter(mAdapter);
                     LinearLayout deleteinterface = findViewById(R.id.deleteinterface);
@@ -301,8 +300,8 @@ public class MainActivity extends AppCompatActivity {
         }
         return super.onOptionsItemSelected(item);
     }
-    public void cancelDelete(View view){
 
+    public void cancelDelete(View view){
         LinearLayout deleteinterface = findViewById(R.id.deleteinterface);
         deleteinterface.setVisibility(View.GONE);
         SlideAnimationUtil.slideInFromTopSlow(getApplicationContext(), findViewById(R.id.atton));
@@ -322,9 +321,8 @@ public class MainActivity extends AppCompatActivity {
         for(int i = 0; i < deletables.size(); i++){
             deleteList = deleteList + deletables.get(i) + ",";
         }
-        System.out.println("ZZZZZZZZZZZZZZ" + deletables);
-        System.out.println("ZZZZZZZZZZZZZZ" + deleteList.substring(0,deleteList.length()-1));
-        db.execSQL("delete from "+FeedReaderContract.FeedEntry.TABLE_NAME+" where "+FeedReaderContract.FeedEntry._ID+" in ("+deleteList.substring(0,deleteList.length()-1)+")");
+        if(!deleteList.isEmpty())
+            db.execSQL("delete from "+FeedReaderContract.FeedEntry.TABLE_NAME+" where "+FeedReaderContract.FeedEntry._ID+" in ("+deleteList.substring(0,deleteList.length()-1)+")");
         db.close();
         SlideAnimationUtil.slideInFromTopSlow(getApplicationContext(), findViewById(R.id.atton));
         SlideAnimationUtil.slideInFromTopSlow(getApplicationContext(), findViewById(R.id.attonu));
@@ -355,7 +353,6 @@ public class MainActivity extends AppCompatActivity {
             strDate = "Yesterday";
         }
         setTitle(strDate);
-        System.out.println(this.getCurrentFocus());
     }
 
     private void SwipeRight() {
@@ -494,7 +491,6 @@ public class MainActivity extends AppCompatActivity {
         values.put(FeedReaderContract.FeedEntry.COLUMN_NAME_STATUS, status);
         values.put(FeedReaderContract.FeedEntry.COLUMN_NAME_DATE, date);
         values.put(FeedReaderContract.FeedEntry.COLUMN_NAME_DATE_MILLI, dateMilli );
-        System.out.println(calendar.getTimeInMillis());
 
         // Insert the new row, returning the primary key value of the new row
         long newRowId = db.insert(FeedReaderContract.FeedEntry.TABLE_NAME, null, values);
@@ -528,6 +524,7 @@ public class MainActivity extends AppCompatActivity {
             refreshTasks();
         }
     }
+
     private boolean getIsOld(){
         Calendar yesterday = Calendar.getInstance();
         yesterday.add(5,-1);
@@ -536,6 +533,7 @@ public class MainActivity extends AppCompatActivity {
         }
         return false;
     }
+
     private void refreshTasks(){
         rollover();
         ArrayList<String> myDataset = Read();
@@ -550,7 +548,6 @@ public class MainActivity extends AppCompatActivity {
             RelativeLayout notaskl = findViewById(R.id.notaskslayout);
             notaskl.setVisibility(View.VISIBLE);
         }
-        System.out.println("ZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZ" + myDataset);
         mAdapter = new MyAdapter(getApplicationContext(),myDataset,myStatus,myID,getIsOld(),false,darkMode);
         recyclerView.setAdapter(mAdapter);
     }
@@ -563,7 +560,6 @@ public class MainActivity extends AppCompatActivity {
             Calendar tempcal = Calendar.getInstance();
             tempcal.set(Calendar.HOUR,0);
             tempcal.set(Calendar.MINUTE,1);
-            Log.d("Rollover Tempcal",tempcal.getTime().toString());
             SimpleDateFormat mdformat = new SimpleDateFormat("MM/dd/yyyy");
             String today =  mdformat.format(tempcal.getTime());
             if(!today.equals(latestdate)){
@@ -600,17 +596,21 @@ public class MainActivity extends AppCompatActivity {
             }
         });
     }
-
+    //hides add task layout on done button click
     public void done(View view) {
         et.setVisibility(View.GONE);
         bt.setVisibility(View.GONE);
         done.setVisibility(View.GONE);
 
-        InputMethodManager inputManager = (InputMethodManager)
-                getSystemService(Context.INPUT_METHOD_SERVICE);
+//        InputMethodManager inputManager = (InputMethodManager)
+//                getSystemService(Context.INPUT_METHOD_SERVICE);
 
-        inputManager.hideSoftInputFromWindow(getCurrentFocus().getWindowToken(),
-                InputMethodManager.HIDE_NOT_ALWAYS);
+//        inputManager.hideSoftInputFromWindow(getCurrentFocus().getWindowToken(),
+//                InputMethodManager.HIDE_NOT_ALWAYS);
+        InputMethodManager manager = (InputMethodManager) this.getSystemService(Context.INPUT_METHOD_SERVICE);
+        if (manager != null) {
+            manager.hideSoftInputFromWindow(this.findViewById(android.R.id.content).getWindowToken(), 0);
+        }
 
         SlideAnimationUtil.slideInFromTopSlow(getApplicationContext(), findViewById(R.id.atton));
         SlideAnimationUtil.slideInFromTopSlow(getApplicationContext(), findViewById(R.id.attonu));
@@ -649,7 +649,6 @@ public class MainActivity extends AppCompatActivity {
             e.printStackTrace();
         }
         String date = jo.toString();
-        Log.d("date for notifications",date);
         editor.putString("notification", date);
         editor.commit();
